Split main file and loader class, passed loader reference to gui
