+++ files/NameIDMap
@@ -9,4 +9,3 @@ snowfall 144
 relativeHumidity 157
 totalCloudCover 164
 solarDuration 46
-
